Start work on reconciler pattern using local queues
